requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
nltk>=3.6.0
PyMuPDF>=1.19.0
spacy>=3.1.0
python-dotenv>=0.19.0
Pillow>=8.3.0
cryptography>=38.0.0
opencv-python>=4.5.0 